refactor: 계층에 따른 역할 분리
