+++ src/service/ProfitService.js
@@ -1,11 +1,11 @@
 import { LOTTO_PRIZE, LOTTO_PRICE } from "../constants/systemConstants.js";
 
-const calculateWinningAmount = (matchingCount) => {
+const calculatePrizeMoney = (matchingCount) => {
   return Object.keys(matchingCount).reduce((sum, count) => sum + matchingCount[count] * (LOTTO_PRIZE[count] || 0), 0);
 };
 
 export const calculateProfitRate = (matchingCount, lottoCount) => {
-  const winningAmount = calculateWinningAmount(matchingCount);
+  const winningAmount = calculatePrizeMoney(matchingCount);
   const profitRatio = winningAmount / (lottoCount * LOTTO_PRICE);
 
   return (profitRatio * 100).toFixed(1);
