+++ src/step1-index.js
@@ -1,63 +1,5 @@
-import { LOTTO_MATCH_CRITERIA } from "./constants/systemConstants.js";
-import SYSTEM_MESSAGE from "./constants/systemMessage.js";
-import WinningLotto from "./domain/WinningLotto.js";
-import { getRetryInput, getBonusNumber, getPrice, getWinningNumber } from "./service/InputService.js";
-import { calculateProfitRate } from "./service/ProfitService.js";
-import { getLottoArray, getLottoCount } from "./service/PurchaseService.js";
-import retryOnError from "./util/retryOnError.js";
-import OutputView from "./view/OutputView.js";
+import GameController from "./controller/GameController.js";
 
-const purchaseLottos = async () => {
-  const price = await retryOnError(getPrice, OutputView.printError);
-  const lottoCount = getLottoCount(price);
-  const lottoArray = getLottoArray(lottoCount);
-  return { lottoArray, lottoCount };
-};
+const game = new GameController();
 
-const makeWinningLotto = async () => {
-  const winningNumbers = await retryOnError(getWinningNumber, OutputView.printError);
-  const bonusNumber = await retryOnError(() => getBonusNumber(winningNumbers), OutputView.printError);
-  const winningLotto = new WinningLotto(winningNumbers, bonusNumber);
-  return winningLotto;
-};
-
-const calculateMatchingResult = (winningLotto, lottoArray) => {
-  const matchingResult = { 3: 0, 4: 0, 5: 0, 6: 0, bonus: 0 };
-
-  lottoArray.forEach((lotto) => {
-    const matchingCount = lotto.match(winningLotto).length;
-
-    if (matchingCount < LOTTO_MATCH_CRITERIA.MIN_MATCH_COUNT) return;
-
-    if (matchingCount === LOTTO_MATCH_CRITERIA.BONUS_MATCH_COUNT && winningLotto.isBonusMatched(lotto)) {
-      matchingResult["bonus"]++;
-      return;
-    }
-
-    matchingResult[matchingCount]++;
-  });
-  return matchingResult;
-};
-
-const runLottoGame = async () => {
-  while (true) {
-    const { lottoArray, lottoCount } = await purchaseLottos();
-
-    OutputView.print(SYSTEM_MESSAGE.COUNT(lottoCount));
-    OutputView.printLottoArray(lottoArray);
-
-    const winningLotto = await makeWinningLotto();
-    const matchingResult = calculateMatchingResult(winningLotto, lottoArray);
-    const profitRate = calculateProfitRate(matchingResult, lottoCount);
-
-    OutputView.printMatchingResult(matchingResult);
-    OutputView.print(SYSTEM_MESSAGE.PROFIT(profitRate));
-
-    const yesOrNo = await retryOnError(getRetryInput, OutputView.printError);
-    if (yesOrNo === "n") {
-      break;
-    }
-  }
-};
-
-await runLottoGame();
+await game.run();
